+++ cohttp-lwt-unix/test/test_sanity_noisy.ml
@@ -45,8 +45,7 @@ let ts_noisy =
         Client.get ~ctx uri >>= fun (resp, body) ->
         assert_equal (Response.status resp) `Not_modified;
         let headers = Response.headers resp in
-        assert_equal ~printer:Transfer.string_of_encoding
-          Transfer.(Fixed 0L)
+        assert_equal ~printer:Transfer.string_of_encoding Transfer.Unknown
           (Header.get_transfer_encoding headers);
         body |> Body.is_empty >|= fun is_empty ->
         assert_bool "No body returned when not modified" is_empty